backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
motor>=3.7.1
pymongo>=4.10.1
certifi>=2025.11.12
pyopenssl>=25.3.0

# LangChain - RAG Orchestration (1.x ecosystem)
langchain>=1.2
langchain-core>=1.2
langchain-community>=0.4
langchain-mongodb>=0.11
langchain-openai>=1.1
langchain-text-splitters>=1.1

# Vector Search & Embeddings
fastembed>=0.7.4

# LLM Integration
openai>=1.59

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
pandas>=3.0.1
openpyxl==3.1.5
python-magic-bin==0.4.14

# Text Processing
tiktoken>=0.8.0

# Rate Limiting
slowapi>=0.1.9

# Utilities
pydantic>=2.12.5
pydantic-settings>=2.13.1
python-dotenv>=1.0.1

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
pytest-cov>=5.0.0
